% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablasDeArticulo.R
\name{generaTablaDescriptivaNumericas}
\alias{generaTablaDescriptivaNumericas}
\title{Crea tabla de resumen descriptivo para variables numéricas}
\usage{
generaTablaDescriptivaNumericas(df, vNumericas, traduccion, columnas = c("n",
  "mediaet", "gauss", "rango", "out3SD", "out5SD"))
}
\arguments{
\item{df}{El data frame}

\item{vNumericas}{El vector que contiene los nombres de las variables numéricas}

\item{Puede}{consistir en repetir el vector vNumericas o bien unos textos que sustituyan el nombre
de cada variable en la tabla de resultado}
}
\value{
Una tabla con las columnas de descriptiva que se indiquen para cada variable numérica
}
\description{
Toma un data frame, una lista de noombres de variables numéricas, una traducción de los nombres de las variables a descripciones más
}
